Improve performance of sub-tabs
This issue was brought up to me by Dragon-seeker <timekeeperguild@gmail.com>. The sub-tabs dropped my FPS from 1000 to 200 in a void world by just opening the menu. Using RenderDoc it is revealed that the sub tabs are being re-rendered unnecessarily in a while-loop, creating around 22,000 draw calls in a single frame. Yikes.

- Draw the sub-tabs during `drawMouseoverTooltip` instead of `renderTabTooltipIfHovered`. This prevents all sub-tabs from re-rendering for each item group.
- Batch the text rendering. This does use a deprecated method which should be present in 1.20.4+, but it makes text render in a single call instead of character-by-character.

Co-authored-by: glisco <glisco03@web.de>